Changed id type

+++ src/main/java/kertz/Supermarket/exception/ArticleNotFoundException.java
@@ -1,7 +1,7 @@
 package kertz.Supermarket.exception;
 
 public class ArticleNotFoundException extends RuntimeException{
-    public ArticleNotFoundException(Integer id){
+    public ArticleNotFoundException(Long id){
         super("Error: Could not found article with id " + id);
     }
 }
